requirements.txt
alembic==1.8.0
aniso8601==9.0.1
apispec==5.2.2
asn1crypto==1.5.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
docopt==0.6.2
Flask==2.1.2
flask-apispec==0.11.1
Flask-Injector==0.14.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Pydantic==0.9.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
injector==0.20.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
mccabe==0.6.1
packaging==21.3
pg8000==1.29.4
pip-api==0.0.29
pipreqs==0.4.11
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.10.4
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
scramp==1.4.4
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.39
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.10
webargs==8.1.0
Werkzeug==2.1.2
yarg==0.1.9
zipp==3.8.0
